Unified both SIDU versions
